requirements.txt
streamlit
openai
ytmusicapi
lyricsgenius
faiss-cpu
python-dotenv
numpy
pandas
seaborn
matplotlib